# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API medico-paciente-app",
        "description": "O objetivo desta API é listar médicos e seus pacientes, divididos por cidade.",
        "contact": {
            "email": "cleison.npi@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Ambiente de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/cidades/{id_cidade}/medicos": {
            "get": {
                "tags": [
                    "cidades x médicos"
                ],
                "summary": "Lista todos os médicos cadastrados de uma cidade",
                "operationId": "74d91fd92e5a69fedf865fabb990fd89",
                "parameters": [
                    {
                        "name": "id_cidade",
                        "in": "path",
                        "description": "ID da cidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna a lista de todos os médicos cadastrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CidadeMedico"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no servidor"
                    }
                }
            }
        },
...
        "/medicos/{id_medico}/pacientes": {
            "get": {
                "tags": [
                    "listar médicos x pacientes"
                ],
                "summary": "Lista todos os pacientes vinculados a um médico",
                "operationId": "f55278c91c448b1497ed831478ef4713",
                "parameters": [
                    {
                        "name": "id_medico",
                        "in": "path",
                        "description": "ID do médico",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Bearer token para autenticação",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna a lista de todos os pacientes vinculados ao médico",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nome": {
                                                        "type": "string",
                                                        "example": "Fábio Gonzales"
                                                    },
                                                    "cpf": {
                                                        "type": "string",
                                                        "example": "795.429.941-60"
                                                    },
                                                    "celular": {
                                                        "type": "string",
                                                        "example": "(11) 98484-6332"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - A autenticação é obrigatória"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "vincular médicos x pacientes"
                ],
                "summary": "Vincula um paciente a um médico",
                "operationId": "40b9ceb20eebc0f3bc8a4cd0ed90eadc",
                "parameters": [
                    {
                        "name": "id_medico",
                        "in": "path",
                        "description": "ID do médico",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Bearer token para autenticação",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "medico_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "paciente_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paciente vinculado ao médico com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "medico": {
                                                    "$ref": "#/components/schemas/Medico"
                                                },
                                                "paciente": {
                                                    "$ref": "#/components/schemas/Paciente"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - A autenticação é obrigatória"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CidadeMedico": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Dra. Alessandra Moura"
                    },
                    "especialidade": {
                        "type": "string",
                        "example": "Neurologista"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "cidade": {
                        "$ref": "#/components/schemas/Cidade"
                    }
                },
                "type": "object"
            },
            "Cidade": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "North Danykamouth"
                    },
                    "estado": {
                        "type": "string",
                        "example": "Wyoming"
                    }
                },
                "type": "object"
            },
            "Medico": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Dra. Alessandra Moura"
                    },
                    "especialidade": {
                        "type": "string",
                        "example": "Neurologista"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "cidade": {
                        "$ref": "#/components/schemas/Cidade"
                    }
                },
                "type": "object"
            },
            "Paciente": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Fábio Gonzales"
                    },
                    "cpf": {
                        "type": "string",
                        "example": "795.429.941-60"
                    },
                    "celular": {
                        "type": "string",
                        "example": "(11) 98484-6332"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}